bb-helper: cancel: Add integration tests
Just converting unit tests to integration tests.
